src/python/requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.9
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
nltk==3.8.1
textblob==0.17.1
plotly==5.18.0
dash==2.14.2
streamlit==1.29.0
sqlalchemy==2.0.23
transformers==4.36.2
torch==2.1.2